Таск и Субтаск пошли получать зарплату, таск получил за зубтаск, а субтаск за таск
